+++ Controllers/TextItemController.cs
@@ -29,7 +29,7 @@ namespace NetflixAPI.Controllers
 
         // GET: api/TextItem/5
         [HttpGet("{text_item_id}")]
-        public async Task<ActionResult<TextItem>> GetTextItem(long id)
+        public async Task<ActionResult<TextItem>> GetTextItem(int id)
         {
             var textItem = await _context.TextItem.FindAsync(id);
 
@@ -44,7 +44,7 @@ namespace NetflixAPI.Controllers
         // PUT: api/TextItem/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{text_item_id}")]
-        public async Task<IActionResult> PutTextItem(long text_item_id, TextItem textItem)
+        public async Task<IActionResult> PutTextItem(int text_item_id, TextItem textItem)
         {
             if (text_item_id != textItem.text_item_id)
             {
@@ -85,7 +85,7 @@ namespace NetflixAPI.Controllers
 
         // DELETE: api/TextItem/5
         [HttpDelete("{text_item_id}")]
-        public async Task<IActionResult> DeleteTextItem(long text_item_id)
+        public async Task<IActionResult> DeleteTextItem(int text_item_id)
         {
             var textItem = await _context.TextItem.FindAsync(text_item_id);
             if (textItem == null)
@@ -99,7 +99,7 @@ namespace NetflixAPI.Controllers
             return NoContent();
         }
 
-        private bool TextItemExists(long text_item_id)
+        private bool TextItemExists(int text_item_id)
         {
             return _context.TextItem.Any(e => e.text_item_id == text_item_id);
         }
