+++ Models/basic/Subscription.cs
@@ -6,10 +6,10 @@ public partial class Subscription
     {
     }
 
-    public long subscription_id {get; set;}
+    public int subscription_id {get; set;}
     public required string subscription_name {get; set;}
 
     public required float subscription_cost {get; set;}
 
-    public virtual required IList<long> quality_id {get; set;}
+    public virtual required IList<int> quality_id {get; set;}
 }
