+++ Models/basic/User.cs
@@ -46,16 +46,16 @@ public partial class User : IValidator
 
     }
 
-    public long user_id {get; set;}
+    public int user_id {get; set;}
     public required string email {get; set;}
     public required string password_hash {get; set;}
     public bool? activated {get; set;}
     public string? blocked_until {get; set;} // string, should probably be datetime (to prevent problems)
-    public long? SubscriptionId {get; set;}
-    public long? language_id {get; set;}
-    public IList<long>? Profiles {get; set;}
-    public IList<long>? FailedLoginAttempts {get; set;}
-    public IList<long>? UserHasInvited {get; set;}
+    public int? SubscriptionId {get; set;}
+    public int? language_id {get; set;}
+    public IList<int>? Profiles {get; set;}
+    public IList<int>? FailedLoginAttempts {get; set;}
+    public IList<int>? UserHasInvited {get; set;}
     public required bool UserHasBeenInvited {get; set;} // should probably not be a Collection (diagram error)
 
 }
