+++ Models/manyToMany/ContentXCLassification.cs
@@ -5,7 +5,7 @@ namespace NetflixAPI.Models;
 [PrimaryKey(nameof(content_id), nameof(classification_id))]
 public class ContentXClassification
 {
-    public long content_id { get; set; }
-    public long classification_id { get; set; }
+    public int content_id { get; set; }
+    public int classification_id { get; set; }
 
 }
